hash functions: add tests to verify output against test vectors
A bug was discovered in the sha256 hash functions provided by this
library but the mistake was not caught for a long time because no tests
were run using known test vectors.

The test vectors used in this commit are from publicly available and
well-respected sources (NIST and the original author of ripe-md).

You can run the tests with "make check".

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>