automake: use libnbcompat.la when linking test files
This ensures that we link with whatever form of libnbcompat autoconf
feels is correct. Note that autoconf makes sure that we are using the
locally-compiled .so as well (in this configuration, the *-test
"binaries" are actually shell scripts that set LD_LIBRARY_PATH among
many other hacks).

I'm not sure whether this change makes any real difference in practice,
but it is probably the way that you're "supposed" to link against a
libtool-ised library within the same project.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>